Introduce env var to set shorter AWS_PROFILE
The AWS_PROFILE var is almost as big as the ARN of the role. Since I
don't want such a long AWS_PROFILE, as most of this info is not really
important for me to look at my starship, I am introducing the
SAML2AWS_PROFILE_SHORT env var. If it is set to true, then a shorter
AWS_PROFILE is set, which is in the format:

account_name/part_of_the_username